fix: separate the backend functionality from the main and format the files
